+++ optimum/neuron/version.py
@@ -12,6 +12,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-__version__ = "0.0.25.dev0"
+__version__ = "0.0.27.dev0"
 
 __sdk_version__ = "2.20.0"
